frontend: switch own-metrics to VictoriaMetrics; refactor metrics client
- Rename frontend/services/prometheus → frontend/services/metrics\n- Point UI to odigos-victoriametrics:8428 (Prometheus-compatible API)\n- Keep PromQL via prometheus/client_golang v1 API\n- Query spans counters for receiver/exporter (accepted/refused with dropped fallback)\n- Update logs/messages; remove old prometheus package files
