+++ frontend/main.go
@@ -31,7 +31,7 @@ import (
 	"github.com/odigos-io/odigos/frontend/services"
 	collectormetrics "github.com/odigos-io/odigos/frontend/services/collector_metrics"
 	"github.com/odigos-io/odigos/frontend/services/db"
-	prommetrics "github.com/odigos-io/odigos/frontend/services/prometheus"
+	metrics "github.com/odigos-io/odigos/frontend/services/metrics"
 	"github.com/odigos-io/odigos/frontend/services/sse"
 	"github.com/odigos-io/odigos/frontend/version"
 	"github.com/odigos-io/odigos/k8sutils/pkg/env"
@@ -63,9 +63,6 @@ var uiFS embed.FS
 
 //go:embed graph/workloads.html
 var workloadsHTML embed.FS
-
-//go:embed graph/odiglet_metrics.html
-var odigletMetricsHTML embed.FS
 
 func parseFlags() Flags {
 	defaultKubeConfig := env.GetDefaultKubeConfigPath()
@@ -255,16 +252,6 @@ func startHTTPServer(ctx context.Context, flags *Flags, logger logr.Logger, odig
 		c.Data(http.StatusOK, "text/html; charset=utf-8", data)
 	})
 
-	// Odiglet collector metrics HTML page
-	r.GET("/odiglet-metrics", func(c *gin.Context) {
-		data, err := odigletMetricsHTML.ReadFile("graph/odiglet_metrics.html")
-		if err != nil {
-			c.String(http.StatusInternalServerError, "Error reading odiglet_metrics.html: %v", err)
-			return
-		}
-		c.Data(http.StatusOK, "text/html; charset=utf-8", data)
-	})
-
 	return r, nil
 }
 
@@ -328,9 +315,9 @@ func main() {
 	}
 
 	var promAPI v1.API
-	promURL := fmt.Sprintf("http://odigos-prometheus.%s.svc:9090", flags.Namespace)
-	if api, err := prommetrics.NewAPIFromURL(promURL); err != nil {
-		log.Printf("Warning: failed to initialize Prometheus API (url=%s): %v", promURL, err)
+	metricsURL := fmt.Sprintf("http://odigos-victoriametrics.%s.svc:8428", flags.Namespace)
+	if api, err := metrics.NewAPIFromURL(metricsURL); err != nil {
+		log.Printf("Warning: failed to initialize VictoriaMetrics API (url=%s): %v", metricsURL, err)
 	} else {
 		promAPI = api
 	}
